[Added] Unproject for camera + [Fix] some warning with CMake, [Added] an operator[] for Vec4f

+++ GalaxyEngine/src/Core/ThreadManager.cpp
@@ -68,7 +68,7 @@ void Core::ThreadManager::Destroy()
 	// clear threadList
 	while (m_threadList.size() > 0)
 	{
-		for (int i = 0; i < m_threadList.size(); i++) {
+		for (size_t i = 0; i < m_threadList.size(); i++) {
 			if (m_threadList[i].joinable())
 			{
 				m_threadList[i].join();
